Create a color ramp

createColorRamp() is a wrapper for GDALCreateColorRamp() in the GDAL API. It automatically creates a color ramp from one color entry to another. Output is an integer matrix in color table format for use with GDALRaster$setColorTable().

Usage

createColorRamp(
  start_index,
  start_color,
  end_index,
  end_color,
  palette_interp = "RGB"
)

Arguments

start_index

Integer start index (raster value).

start_color

Integer vector of length three or four. A color entry value to start the ramp (e.g., RGB values).

end_index

Integer end index (raster value).

end_color

Integer vector of length three or four. A color entry value to end the ramp (e.g., RGB values).

palette_interp

One of "Gray", "RGB" (the default), "CMYK" or "HLS" descibing interpretation of start_color and end_color values (see GDAL Color Table).

Value

Intger matrix with five columns containing the color ramp from start_index to end_index, with raster index values in column 1 and color entries in columns 2:5).

Note

createColorRamp() could be called several times, using rbind() to combine multiple ramps into the same color table. Possible duplicate rows in the resulting table are not a problem when used in GDALRaster$setColorTable() (i.e., when end_color of one ramp is the same as start_color of the next ramp).

See also

Examples

# create a color ramp for tree canopy cover percent
# band 5 of an LCP file contains canopy cover
lcp_file <- system.file("extdata/storm_lake.lcp", package="gdalraster")
ds <- new(GDALRaster, lcp_file, read_only=TRUE)
ds$getDescription(band=5)
#> [1] "Canopy cover"
ds$getMetadata(band=5, domain="")
#> [1] "CANOPY_COV_FILE=/netapp/sharedwebfs1/shared/landfire/public/temp_q8dTbIJ4w6Qi36Omkzk0/LCP_LF2022_FBFM40_220_CONUS/temp/merged_modified.tif"
#> [2] "CANOPY_COV_MAX=75"                                                                                                                         
#> [3] "CANOPY_COV_MIN=-9999"                                                                                                                      
#> [4] "CANOPY_COV_NUM_CLASSES=8"                                                                                                                  
#> [5] "CANOPY_COV_UNIT=1"                                                                                                                         
#> [6] "CANOPY_COV_UNIT_NAME=Percent"                                                                                                              
ds$close()

# create a GTiff file with Byte data type for the canopy cover band
# recode nodata -9999 to 255
tcc_file <- calc(expr = "ifelse(CANCOV == -9999, 255, CANCOV)",
                 rasterfiles = lcp_file,
                 bands = 5,
                 var.names = "CANCOV",
                 fmt = "GTiff",
                 dtName = "Byte",
                 nodata_value = 255,
                 setRasterNodataValue = TRUE)
#> Error in calc(expr = "ifelse(CANCOV == -9999, 255, CANCOV)", rasterfiles = lcp_file,     bands = 5, var.names = "CANCOV", fmt = "GTiff", dtName = "Byte",     nodata_value = 255, setRasterNodataValue = TRUE): could not find function "calc"

ds_tcc <- new(GDALRaster, tcc_file, read_only=FALSE)
#> Error in eval(expr, envir, enclos): object 'tcc_file' not found

# create a color ramp from 0 to 100 and set as the color table
colors <- createColorRamp(start_index = 0,
                          start_color = c(211, 211, 211),
                          end_index = 100,
                          end_color = c(0, 100, 0))
#> Error in createColorRamp(start_index = 0, start_color = c(211, 211, 211),     end_index = 100, end_color = c(0, 100, 0)): could not find function "createColorRamp"

print(colors)
#> function (distinct = FALSE) 
#> {
#>     c <- .Call(C_colors)
#>     if (distinct) 
#>         c[!duplicated(t(col2rgb(c)))]
#>     else c
#> }
#> <bytecode: 0x562a7f497c18>
#> <environment: namespace:grDevices>
ds_tcc$setColorTable(band=1, col_tbl=colors, palette_interp="RGB")
#> Error in eval(expr, envir, enclos): object 'ds_tcc' not found
ds_tcc$setRasterColorInterp(band=1, col_interp="Palette")
#> Error in eval(expr, envir, enclos): object 'ds_tcc' not found

# close and re-open the dataset in read_only mode
ds_tcc$open(read_only=TRUE)
#> Error in eval(expr, envir, enclos): object 'ds_tcc' not found

plot_raster(ds_tcc, interpolate=FALSE, legend=TRUE,
            main="Storm Lake Tree Canopy Cover (%)")
#> Error in eval(expr, envir, enclos): object 'ds_tcc' not found
ds_tcc$close()
#> Error in eval(expr, envir, enclos): object 'ds_tcc' not found