Get the size of memory in use by the GDAL block cache

get_cache_used() returns the amount of memory currently in use for GDAL block caching. This a wrapper for GDALGetCacheUsed64() with return value as MB.

Usage

get_cache_used()

Value

Integer. Amount of cache memory in use in MB.

See also

Examples

get_cache_used()
#> [1] 0