Display raster data that has been read into a vector

plot_raster_data() displays raster data using base graphics.

Usage

plot_raster_data(
  data,
  xsize,
  ysize,
  nbands = 1,
  col_tbl = NULL,
  normalize = TRUE,
  col_map_fn = NULL,
  xlim = c(0, xsize),
  ylim = c(ysize, 0),
  xlab = "x",
  ylab = "y",
  interpolate = TRUE,
  asp = 1,
  axes = TRUE,
  main = "",
  xaxs = "i",
  yaxs = "i",
  na_col = rgb(0, 0, 0, 0),
  ...
)

Arguments

data

A numeric vector of pixel data to display, arranged in left to right, top to bottom pixel order.

xsize

The number of pixels along the x dimension in data.

ysize

The number of pixels along the y dimension in data.

nbands

The number of bands in data. Must be either 1 (grayscale) or 3 (RGB). For RGB, data are interleaved by band.

col_tbl

A color table as a data frame with four columns. Column 1 contains the integer raster values, columns 2:4 contain the intensities (between 0 and 1) of the red, green and blue primaries.

normalize

Logical. TRUE to rescale pixel values so that their range is [0,1]. Ignored if col_tbl is used.

col_map_fn

An optional color map function (default is grDevices::gray for single-band data or grDevices::rgb for 3-band). Ignored if col_tbl is used.

xlim

Numeric vector of length two giving the x coordinate range. The default uses pixel/line space (c(0, xsize)).

ylim

Numeric vector of length two giving the y coordinate range. The default uses pixel/line space (c(ysize, 0)).

xlab

Title for the x axis (see ?title).

ylab

Title for the y axis (see ?title).

interpolate

Logical indicating whether to apply linear interpolation to the image when drawing (default TRUE).

asp

Numeric. The aspect ratio y/x (see ?plot.window).

axes

Logical. TRUE to draw axes (the default).

main

The main title (on top).

xaxs

The style of axis interval calculation to be used for the x axis (see ?par).

yaxs

The style of axis interval calculation to be used for the y axis (see ?par).

na_col

Color to use for NA as a 7- or 9-character hexadecimal code. The default is transparent ("#00000000", the return value of rgb(0,0,0,0)).

...

Other parameters to be passed to plot.default().

Examples

## Elevation
elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
ds <- new(GDALRaster, elev_file, read_only=TRUE)
ncols <- ds$getRasterXSize()
nrows <- ds$getRasterYSize()
r <- read_ds(ds)
ds$close()

# grayscale
plot_raster_data(r, xsize=ncols, ysize=nrows, main="Storm Lake elevation")
#> Error in if (grDevices::dev.capabilities()$rasterImage == "no") {    message("Device does not support rasterImage().")    invisible(FALSE)}: missing value where TRUE/FALSE needed

# color ramp using 'terrain.colors' in base R
col_ramp <- scales::gradient_n_pal(terrain.colors(10))
plot_raster_data(r, xsize=ncols, ysize=nrows, col_map_fn=col_ramp,
                 main="Storm Lake: elevation")
#> Error in if (grDevices::dev.capabilities()$rasterImage == "no") {    message("Device does not support rasterImage().")    invisible(FALSE)}: missing value where TRUE/FALSE needed

## Landsat band combination
b4_file <- system.file("extdata/sr_b4_20200829.tif", package="gdalraster")
b5_file <- system.file("extdata/sr_b5_20200829.tif", package="gdalraster")
b6_file <- system.file("extdata/sr_b6_20200829.tif", package="gdalraster")
band_files <- c(b6_file, b5_file, b4_file)

ds <- new(GDALRaster, b4_file, read_only=TRUE)
ncols <- ds$getRasterXSize()
nrows <- ds$getRasterYSize()
ds$close()

r <- vector("integer")
for (f in band_files) {
  ds <- new(GDALRaster, f, read_only=TRUE)
  r <- c(r, read_ds(ds))
  ds$close()
}

plot_raster_data(r, xsize=ncols, ysize=nrows, nbands=3,
                 main="Storm Lake: Landsat 6-5-4 (vegetative analysis)")
#> Error in if (grDevices::dev.capabilities()$rasterImage == "no") {    message("Device does not support rasterImage().")    invisible(FALSE)}: missing value where TRUE/FALSE needed

## LANDFIRE vegetation with colors from the CSV attribute table
evc_file <- system.file("extdata/storml_evc.tif", package="gdalraster")
evc_vat <- system.file("extdata/LF20_EVC_220.csv", package="gdalraster")
vat <- read.csv(evc_vat)
head(vat)
#>   VALUE                        CLASSNAMES   R   G   B      RED    GREEN
#> 1 -9999                       Fill-NoData 255 255 255 1.000000 1.000000
#> 2    11                        Open Water   0   0 255 0.000000 0.000000
#> 3    12                          Snow/Ice 159 161 240 0.623529 0.631373
#> 4    13 Developed-Upland Deciduous Forest  64  61 168 0.250980 0.239216
#> 5    14 Developed-Upland Evergreen Forest  68  79 137 0.266667 0.309804
#> 6    15     Developed-Upland Mixed Forest 102 119 205 0.400000 0.466667
#>       BLUE
#> 1 1.000000
#> 2 1.000000
#> 3 0.941176
#> 4 0.658824
#> 5 0.537255
#> 6 0.803922

ds <- new(GDALRaster, evc_file, read_only=TRUE)
ncols <- ds$getRasterXSize()
nrows <- ds$getRasterYSize()

r <- read_ds(ds)
ds$close()
plot_raster_data(r, xsize=ncols, ysize=nrows, col_tbl=vat[,c(1,6:8)],
                 main="Storm Lake: LANDFIRE canopy cover")
#> Error in if (grDevices::dev.capabilities()$rasterImage == "no") {    message("Device does not support rasterImage().")    invisible(FALSE)}: missing value where TRUE/FALSE needed