Obtain information about coordinate reference systems in the PROJ DB

srs_info_from_db() returns a data frame containing descriptive information about spatial coordinate reference systems in the PROJ database. Wrapper of OSRGetCRSInfoListFromDatabase() in the GDAL SRS API.

Usage

srs_info_from_db(auth_name = "")

Arguments

auth_name

Character string containing an authority name used to restrict the search, or empty string ("") for all authorities (the default).

Value

A data frame (will contain 0 rows if no objects are found that match the search criterion).

Details

The returned information includes the authority name, object code, object name, object type (e.g., Geographic 2D CRS, Geographic 3D CRS, Projected CRS, etc.), whether the object is deprecated, whether the bounding box values for the area of use are valid, the bounding box values in longitude and latitude degrees, the name of the area of use, the name of the projection method for a projected CRS, and the name of the celestial body of the CRS (e.g., "Earth", populated only if GDAL >= 3.12 and PROJ >= 8.1).

See also

Examples

epsg <- srs_info_from_db("EPSG")
str(epsg)
#> 'data.frame':	7497 obs. of  13 variables:
#>  $ auth_name          : chr  "EPSG" "EPSG" "EPSG" "EPSG" ...
#>  $ obj_code           : chr  "2000" "2001" "2002" "2003" ...
#>  $ obj_name           : chr  "Anguilla 1957 / British West Indies Grid" "Antigua 1943 / British West Indies Grid" "Dominica 1945 / British West Indies Grid" "Grenada 1953 / British West Indies Grid" ...
#>  $ obj_type           : chr  "Projected" "Projected" "Projected" "Projected" ...
#>  $ is_deprecated      : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ is_bbox_valid      : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ west_long_deg      : num  -63.2 -62 -61.5 -61.8 -62.3 ...
#>  $ south_lat_deg      : num  18.1 16.9 15.1 11.9 16.6 ...
#>  $ east_long_deg      : num  -62.9 -61.6 -61.2 -61.4 -62.1 ...
#>  $ north_lat_deg      : num  18.3 17.2 15.7 12.6 16.9 ...
#>  $ area_of_use        : chr  "Anguilla - onshore." "Antigua island - onshore." "Dominica - onshore." "Grenada and southern Grenadine Islands - onshore." ...
#>  $ projection_method  : chr  "Transverse Mercator" "Transverse Mercator" "Transverse Mercator" "Transverse Mercator" ...
#>  $ celestial_body_name: chr  NA NA NA NA ...

iau <- srs_info_from_db("IAU_2015")
str(iau)
#> 'data.frame':	2201 obs. of  13 variables:
#>  $ auth_name          : chr  "IAU_2015" "IAU_2015" "IAU_2015" "IAU_2015" ...
#>  $ obj_code           : chr  "1000" "1010" "1015" "1020" ...
#>  $ obj_name           : chr  "Sun (2015) - Sphere / Ocentric" "Sun (2015) - Sphere / Ocentric / Equirectangular, clon = 0" "Sun (2015) - Sphere / Ocentric / Equirectangular, clon = 180" "Sun (2015) - Sphere / Ocentric / Sinusoidal, clon = 0" ...
#>  $ obj_type           : chr  "Geographic 2D" "Projected" "Projected" "Projected" ...
#>  $ is_deprecated      : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ is_bbox_valid      : logi  TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ west_long_deg      : num  -180 -180 -180 -180 -180 -180 -180 -180 -180 -180 ...
#>  $ south_lat_deg      : num  -90 -90 -90 -90 -90 -90 -90 -90 -90 -90 ...
#>  $ east_long_deg      : num  180 180 180 180 180 180 180 180 180 180 ...
#>  $ north_lat_deg      : num  90 90 90 90 90 90 90 90 90 90 ...
#>  $ area_of_use        : chr  "Not specified." "Not specified." "Not specified." "Not specified." ...
#>  $ projection_method  : chr  NA "Equidistant Cylindrical" "Equidistant Cylindrical" "Sinusoidal" ...
#>  $ celestial_body_name: chr  NA NA NA NA ...