Raster pixel/line from geospatial x,y coordinates

get_pixel_line() converts geospatial coordinates to pixel/line (raster column, row numbers). The upper left corner pixel is the raster origin (0,0) with column, row increasing left to right, top to bottom.

Usage

get_pixel_line(xy, gt)

Arguments

xy

Numeric array of geospatial x,y coordinates in the same spatial reference system as gt.

gt

Numeric vector of length six. The affine geotransform for the raster.

Value

Integer array of raster pixel/line.

See also

Examples

pt_file <- system.file("extdata/storml_pts.csv", package="gdalraster")
## id, x, y in NAD83 / UTM zone 12N
pts <- read.csv(pt_file)
print(pts)
#>    id   xcoord  ycoord
#> 1   1 324650.9 5103344
#> 2   2 324171.0 5103034
#> 3   3 323533.4 5103329
#> 4   4 325220.0 5103508
#> 5   5 325703.1 5102377
#> 6   6 326297.8 5103924
#> 7   7 325520.4 5104146
#> 8   8 326247.7 5102506
#> 9   9 327711.7 5104476
#> 10 10 324181.7 5103901
raster_file <- system.file("extdata/storm_lake.lcp", package="gdalraster")
ds <- new(GDALRaster, raster_file)
gt <- ds$getGeoTransform()
get_pixel_line(as.matrix(pts[,-1]), gt)
#>       [,1] [,2]
#>  [1,]   39   57
#>  [2,]   23   68
#>  [3,]    1   58
#>  [4,]   58   52
#>  [5,]   74   90
#>  [6,]   94   38
#>  [7,]   68   31
#>  [8,]   92   85
#>  [9,]  141   20
#> [10,]   23   39
ds$close()